justfile: add gen-admin-macs command
move MuliHost to enums.py
